day10/day10_test.go
package main

import (
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestNewCPU(t *testing.T) {
	cpu := newCpu()
	expected_cpu := CPU{X: 1, Cycles: 1}

	assert.Equal(t, cpu, expected_cpu)
}

func TestGetSignalStrengthSum(t *testing.T) {
	cpu := newCpu()
	result := cpu.getSignalStrengthSum(sampleProgram())
	assert.Equal(t, result, 13140)
}

func TestNoopIncreasesCycles(t *testing.T) {
	cpu := CPU{X: 10, Cycles: 10}
	cpu.noop()

	assert.Equal(t, 10, cpu.X)
	assert.Equal(t, 11, cpu.Cycles)
}

func TestNoopLogsKeyNumber(t *testing.T) {
	cpu := CPU{X: 5, Cycles: 19}
	cpu.noop()

	assert.Equal(t, 5, cpu.X)
	assert.Equal(t, 20, cpu.Cycles)
	assert.Equal(t, 100, cpu.Total)
}

func TestAddxIncreasesCycles(t *testing.T) {
	cpu := CPU{X: 75, Cycles: 4}
	cpu.addx("25")

	assert.Equal(t, 100, cpu.X)
	assert.Equal(t, 6, cpu.Cycles)
}

func sampleProgram() []string {
	return []string{
		"addx 15",
		"addx -11",
		"addx 6",
		"addx -3",
		"addx 5",
		"addx -1",
		"addx -8",
		"addx 13",
		"addx 4",
		"noop",
		"addx -1",
		"addx 5",
		"addx -1",
		"addx 5",
		"addx -1",
		"addx 5",
		"addx -1",
		"addx 5",
		"addx -1",
		"addx -35",
		"addx 1",
		"addx 24",
		"addx -19",
		"addx 1",
		"addx 16",
		"addx -11",
		"noop",
		"noop",
		"addx 21",
		"addx -15",
		"noop",
		"noop",
		"addx -3",
		"addx 9",
		"addx 1",
		"addx -3",
		"addx 8",
		"addx 1",
		"addx 5",
		"noop",
		"noop",
		"noop",
		"noop",
		"noop",
		"addx -36",
		"noop",
		"addx 1",
		"addx 7",
		"noop",
		"noop",
		"noop",
		"addx 2",
		"addx 6",
		"noop",
		"noop",
		"noop",
		"noop",
		"noop",
		"addx 1",
		"noop",
		"noop",
		"addx 7",
		"addx 1",
		"noop",
		"addx -13",
		"addx 13",
		"addx 7",
		"noop",
		"addx 1",
		"addx -33",
		"noop",
		"noop",
		"noop",
		"addx 2",
		"noop",
		"noop",
		"noop",
		"addx 8",
		"noop",
		"addx -1",
		"addx 2",
		"addx 1",
		"noop",
		"addx 17",
		"addx -9",
		"addx 1",
		"addx 1",
		"addx -3",
		"addx 11",
		"noop",
		"noop",
		"addx 1",
		"noop",
		"addx 1",
		"noop",
		"noop",
		"addx -13",
		"addx -19",
		"addx 1",
		"addx 3",
		"addx 26",
		"addx -30",
		"addx 12",
		"addx -1",
		"addx 3",
		"addx 1",
		"noop",
		"noop",
		"noop",
		"addx -9",
		"addx 18",
		"addx 1",
		"addx 2",
		"noop",
		"noop",
		"addx 9",
		"noop",
		"noop",
		"noop",
		"addx -1",
		"addx 2",
		"addx -37",
		"addx 1",
		"addx 3",
		"noop",
		"addx 15",
		"addx -21",
		"addx 22",
		"addx -6",
		"addx 1",
		"noop",
		"addx 2",
		"addx 1",
		"noop",
		"addx -10",
		"noop",
		"noop",
		"addx 20",
		"addx 1",
		"addx 2",
		"addx 2",
		"addx -6",
		"addx -11",
		"noop",
		"noop",
		"noop",
	}
}
